情感之家最新版本提交2020.9.30

+++ app/src/main/java/com/yc/emotion/home/mine/ui/activity/MyTestReportActivity.kt
@@ -53,7 +53,7 @@ class MyTestReportActivity : BaseSameActivity(), EmotionTestView {
         initData()
         initListener()
 
-        userId = UserInfoHelper.instance.getUid() as Int
+        userId = UserInfoHelper.instance.getUid()
 
 
     }
